Fix Mms notification not always clearing.
This is a two fold commit based on work done by Ian Ling to CarbonRom.

https://github.com/CarbonDev/android_packages_apps_Mms/commit/b3a831a6e3019d256163b23870e0f4e0134fce16

Clear notification when tapped
Change-Id: I1e660088436e36efdf4bfb000f273b63f5462ae0

https://github.com/CarbonDev/android_packages_apps_Mms/commit/21c851e139c56f387c83e11d12d2a2fa8e0c3ff0

Mms: Mark messages read if they are on-screen when phone is unlocked
Previous behavior: Open Messaging app, open one of the existing conversations,
and then lock the phone. Receive a new text message in that conversation.
Unlock the phone to reveal the new message, but the notification does not
disappear despite the new message having been read.

This is due to onResume() not getting called read when returning to
the app from the lockscreen, so the new message never gets marked as read!

The solution is to use onWindowFocusChanged() as a failsafe for this scenario,
as this DOES get called when you come in from the lockscreen. Now the notification
disappears correctly.

Change-Id: I225a31bea9abe553c1466e070f4c9b50245f05c5
Signed-off-by: baked <travp624@gmail.com>